[bug]티켓페이지 디자인 수정
